Fix: Fix with copy public directory errors of issue.

+++ src/command/build/handlerBuild.ts
@@ -23,6 +23,7 @@ const WARN_AFTER_BUNDLE_GZIP_SIZE = 512 * 1024;
 const WARN_AFTER_CHUNK_GZIP_SIZE = 1024 * 1024;
 
 function copyPublicFolder() {
+  if (!fs.existsSync(paths.appPublic)) return;
   fs.copySync(paths.appPublic, paths.appBuild, {
     dereference: true,
     filter: file => file !== paths.appHtml,
